Feat - automatic testing (#61)
* Automatic testing

Add basic functionality for automatic testing. Runs a script which
runs coverage with pytest.

Closes #6

* Add test

Just to see that tests are actually ran

Closes #32

* fix - created tests for #61

* fix - remove unused import #61

* feat - navigate build history #63 (#65)

* Fix - update database table (#67)

* UPDATE database

Switch out build_log field for two fields, linter and test results.

Fixes #64

* Declutter

Remove some code that had been left in by mistake.

---------

Co-authored-by: jacmol@kth.se <jacmol@kth.se>

* Feat/webhook (#66)

adding functionality to handle the webhook in one file instead of 3.
deleted unnecessary router files.
Handling payloads and can now clone, and check the syntax of the github repo that was pushed.
closes #44

* Fix - Add query wrapper (#70)

* Add query wrapper

To check if the database exists before performing query

Fixes #68

* fix - switch db api to use orm

---------

Co-authored-by: jacmol@kth.se <jacmol@kth.se>
Co-authored-by: Samuel Söderberg <samuelsoderberg99@gmail.com>

* fix/handle multiple commits (#72)

Handles the case when there are multiple commits. It splits each clone into a separate folder based on push id and then after running the syntax check it will delete the repo. Testing also implemented.

closes #71

* Feat/branch (#73)

Clones the repository, checks out the branch specified in the payload, and pulls from this branch to run the syntax check and tests. 

closes #27

---------

Co-authored-by: jacmol@kth.se <jacmol@kth.se>
Co-authored-by: Samuel Söderberg <samuelsoderberg99@gmail.com>
Co-authored-by: Phoebe Schwartz <144394710+peschwartz@users.noreply.github.com>